+++ app/Plugin/Sales/Model/ProductSpecificationPart.php
@@ -38,30 +38,46 @@ class ProductSpecificationPart extends AppModel {
 
 	public function savePart($partdata = null , $auth = null,$specId = null){
 		$Ids = array();
-		
-		foreach ($partdata[$this->name] as $key => $partList) {
-			$this->create();
-			$partList['created_by'] = $auth;
-			$partList['modified_by'] = $auth;
-			$partList['product_specification_id'] = $specId;
-			$partList['product_id'] = $partdata['Product']['id'];
 			
-			$this->save($partList);
-			if (!empty($partList['order'])) {
-				
-			array_push($Ids, $this->id.'-'.$partList['order'].'-'.'Part');
-			}
+			if (!empty($partdata[$this->name])) {
+
+				foreach ($partdata[$this->name] as $key => $partList) {
+					$this->create();
+					$partList['created_by'] = $auth;
+					$partList['modified_by'] = $auth;
+					$partList['product_specification_id'] = $specId;
+					$partList['product_id'] = $partdata['Product']['id'];
+					
+					$this->save($partList);
+						
+					array_push($Ids, $this->id.'-'.$partList['order'].'-'.'Part');
+				}
+
 		}
+
 		return $Ids;
 
 	}
 
 	public function deleteData($partData = null){
 
-		foreach ($partData['ProductSpecificationPart'] as $key => $deleteMe) {
-			$this->delete($deleteMe);
-		}
+		if (!empty($partData['ProductSpecificationPart']) && is_array($partData['ProductSpecificationPart'])) {
+			
+			foreach ($partData['ProductSpecificationPart'] as $key => $deleteMe) {
 
+				if (is_array($deleteMe)) {
+
+					foreach ($deleteMe as $key => $innerDelete) {
+						$this->delete($innerDelete);
+					}
+
+				} else {
+					
+					$this->delete($deleteMe);
+				}
+			}
+		}
+		
 	}
 
 }
